requirements.txt
crewai
crewai_tools
crewai_google_genai
python-dotenv
langchain_google_genai